refactor: split header action row into its own component
