+++ client/src/components/common/ArtistHeaderSection.tsx
@@ -1,22 +1,18 @@
 import { css } from "@emotion/css";
 import styled from "@emotion/styled";
-import { ArtistButtonLink } from "components/Artist/ArtistButtons";
-import ArtistHeaderActionsStrip from "components/Artist/ArtistHeaderActionsStrip";
 import Avatar from "components/Artist/Avatar";
 import LoadingBlocks from "components/Artist/LoadingBlocks";
 import { MetaCard } from "components/common/MetaCard";
-import ArtistFormLabel from "components/ManageArtist/ArtistFormLabel";
 import ArtistFormLocation from "components/ManageArtist/ArtistFormLocation";
 import { UpdateArtistBody, useUpdateArtistMutation } from "queries";
 import React from "react";
-import { useTranslation } from "react-i18next";
 import { useAuthContext } from "state/AuthContext";
 import { useSnackbar } from "state/SnackbarContext";
-import { getArtistUrl } from "utils/artist";
 import { useFitTitle } from "utils/useFitTitle";
 
 import { between, bp } from "../../constants";
 
+import ArtistHeaderActionsRow from "./ArtistHeaderActionsRow";
 import FollowArtist from "./FollowArtist";
 import SpaceBetweenDiv from "./SpaceBetweenDiv";
 
@@ -134,23 +130,11 @@ const ArtistHeaderSection: React.FC<{
   isManage: boolean;
 }> = ({ artist, isLoading, isManage }) => {
   const artistAvatar = artist?.avatar;
-  const { t } = useTranslation("translation", { keyPrefix: "manageArtist" });
-
-  const { t: tArtist } = useTranslation("translation", { keyPrefix: "artist" });
 
   const { user } = useAuthContext();
   const { mutateAsync: updateArtist } = useUpdateArtistMutation();
   const snackbar = useSnackbar();
   const titleRef = useFitTitle<HTMLHeadingElement>({ deps: [artist?.name] });
-
-  const displayedLabelArtist = React.useMemo(() => {
-    if (!artist || artist.isLabelProfile) return null;
-    const match = (artist.artistLabels ?? []).find(
-      (al) =>
-        al.isDisplayedOnArtistPage && al.isArtistApproved && al.isLabelApproved
-    );
-    return match?.labelUser.artists?.[0] ?? null;
-  }, [artist]);
 
   const handleSubmit = React.useCallback(
     async (data: UpdateArtistBody) => {
@@ -271,43 +255,11 @@ const ArtistHeaderSection: React.FC<{
             `}
           />
         )}
-        <div
-          className={`w-full flex flex-row items-baseline justify-end pb-2 md:justify-between max-md:py-1 max-md:border-t max-md:border-(--mi-button-color)/50 ${displayedLabelArtist ? "md:mt-2" : "md:mt-1"}`}
-        >
-          <div className="flex flex-row items-baseline gap-2">
-            {displayedLabelArtist && (
-              <ArtistButtonLink
-                variant="chip"
-                to={getArtistUrl(displayedLabelArtist)}
-                className={css`
-                  &[class] {
-                    background-color: color-mix(
-                      in srgb,
-                      var(--mi-button-color) 8%,
-                      transparent
-                    ) !important;
-                    filter: brightness(var(--mi-chip-brightness, 1));
-                  }
-                  @media screen and (max-width: ${bp.medium}px) {
-                    &[class] {
-                      display: none !important;
-                    }
-                  }
-                `}
-              >
-                {tArtist("onLabel", {
-                  name: displayedLabelArtist.name,
-                })}
-              </ArtistButtonLink>
-            )}
-            {isManage && <ArtistFormLabel artist={artist} />}
-          </div>
-          <ArtistHeaderActionsStrip
-            artist={artist}
-            isManage={!!isManage}
-            onSubmit={handleSubmit}
-          />
-        </div>
+        <ArtistHeaderActionsRow
+          artist={artist}
+          isManage={!!isManage}
+          onSubmit={handleSubmit}
+        />
       </HeaderWrapper>
     </div>
   );
